inclusive-education-system/ai-services/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
mysql-connector-python==8.0.33
gunicorn==21.2.0
PyPDF2==3.0.1
pymysql==1.1.0
